Port wallet breakdown to vue
