+++ index.template.html
@@ -122,24 +122,6 @@
           </div>
         </div>
 
-        <!-- // Wallet Breakdown -->
-        <div class="modal fade" id="walletBreakdownModal" tabindex="-1" aria-labelledby="walletBreakdownModalLabel" role="dialog" aria-hidden="true">
-          <div class="modal-dialog modal-dialog-centered" role="document">
-            <div class="modal-content exportKeysModalColor">
-              <div class="modal-header" style="display: flex; justify-content: center; padding-bottom: 0px;">
-                <h5 data-i18n="balanceBreakdown" class="modal-title" id="walletBreakdownModalLabel">Balance Breakdown</h5>
-              </div>
-              <div class="modal-body center-text" style="padding-bottom: 0px; padding-top: 0px;">
-                <canvas id="walletBreakdownCanvas"></canvas>
-                <div id="walletBreakdownLegend" style="padding-left: 14px; padding-top: 20px;"></div>
-              </div>
-              <div class="modal-footer" style="padding-top: 0px; display: flex; justify-content: center;">
-                <button type="button" class="pivx-button-big-cancel" data-dismiss="modal" aria-label="Close">Close</button>
-              </div>
-            </div>
-          </div>
-        </div>
-
         <br />
         <!-- // CONFIRM MODAL -->
         <div class="modal" id="confirmModal" tabindex="-1" style="width: 100%; z-index: 2000; background-color: #201436db;" role="dialog" aria-hidden="true" data-backdrop="static" data-keyboard="false">
